moved DashboardLayout component to components folder
