+++ front-end/src/components/DashboardLayout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import SideNavBar from "../components/SideNavBar";
+import SideNavBar from "./SideNavBar";
 
 export default function DashboardLayout() {
   return (
